@@ -1,5 +1,5 @@
 {
-    "name": "cakasim/payone-sdk-http-message",
+    "name": "andrepayone/payone-sdk-http-message",
     "description": "A PSR-7 and PSR-17 Implementation for the SDK for PAYONE Payment Integration.",
     "type": "library",
     "keywords": [
@@ -9,7 +9,7 @@
         "psr-7",
         "psr-17"
     ],
-    "homepage": "https://cakasim.de",
+    "homepage": "https://payone.de",
     "readme": "README.md",
     "license": "MIT",
     "authors": [
@@ -18,33 +18,39 @@
             "email": "me@cakasim.de",
             "homepage": "https://cakasim.de",
             "role": "Developer"
+        },
+        {
+            "name": "André Gregor-Herrmann",
+            "email": "andre.gregor-herrmann@payone.com",
+            "homepage": "https://payone.de",
+            "role": "Developer"
         }
     ],
     "support": {
         "email": "support@cakasim.de",
-        "issues": "https://github.com/Cakasim/php-payone-sdk-http-message/issues",
-        "source": "https://github.com/Cakasim/php-payone-sdk-http-message"
+        "issues": "https://github.com/andrepayone/php-payone-sdk-http-message/issues",
+        "source": "https://github.com/andrepayone/php-payone-sdk-http-message"
     },
     "minimum-stability": "stable",
     "require": {
-        "php": "^7.1",
+        "php": "^8.1",
         "psr/http-message": "^1.0",
         "psr/http-factory": "^1.0"
     },
     "autoload": {
         "psr-4": {
-            "Cakasim\\Payone\\Sdk\\Http\\": "src/"
+            "Payone\\Sdk\\Http\\": "src/"
         }
     },
     "autoload-dev": {
         "psr-4": {
-            "Cakasim\\Payone\\Sdk\\Tests\\Http\\": "tests/"
+            "Payone\\Sdk\\Tests\\Http\\": "tests/"
         }
     },
     "require-dev": {},
     "scripts": {
-        "download-phpcsf": "curl --silent -L https://cs.symfony.com/download/php-cs-fixer-v2.phar --create-dirs --output vendor/bin/phpcsf.phar",
-        "download-phpstan": "curl --silent -L https://github.com/phpstan/phpstan/releases/download/0.12.19/phpstan.phar --create-dirs --output vendor/bin/phpstan.phar",
+        "download-phpcsf": "curl --silent -L https://cs.symfony.com/download/php-cs-fixer-v3.phar --create-dirs --output vendor/bin/phpcsf.phar",
+        "download-phpstan": "curl --silent -L https://github.com/phpstan/phpstan/releases/download/1.10.38/phpstan.phar --create-dirs --output vendor/bin/phpstan.phar",
         "download-phpunit": "curl --silent -L https://phar.phpunit.de/phpunit-8.phar --create-dirs --output vendor/bin/phpunit.phar",
         "phpcsf": "php vendor/bin/phpcsf.phar -v fix --dry-run --allow-risky=yes",
         "phpstan": "php vendor/bin/phpstan.phar analyse",
