@@ -2,10 +2,10 @@
 
 declare(strict_types=1);
 
-namespace Cakasim\Payone\Sdk\Tests\Http\Message;
+namespace Payone\Sdk\Tests\Http\Message;
 
-use Cakasim\Payone\Sdk\Http\Factory\ResponseFactory;
-use Cakasim\Payone\Sdk\Http\Factory\StreamFactory;
+use Payone\Sdk\Http\Factory\ResponseFactory;
+use Payone\Sdk\Http\Factory\StreamFactory;
 use PHPUnit\Framework\TestCase;
 use Psr\Http\Message\ResponseInterface;
 
